--- a/package.json
+++ b/package.json
@@ -305,6 +305,7 @@
     "fresnel-17": "npm:@artsy/fresnel@6",
     "graphql-tag": "2.10.3",
     "graphql-tools": "4.0.3",
+    "html-webpack-plugin": "^5.6.3",
     "hulk-editor": "craigspaeth/hulk",
     "husky": "3.0.5",
     "immer": "9.0.7",