--- a/webpack/sharedPlugins.js
+++ b/webpack/sharedPlugins.js
@@ -3,23 +3,23 @@
 import { RetryChunkLoadPlugin } from "webpack-retry-chunk-load-plugin"
 import { EarlyHintsPlugin } from "./plugins/EarlyHintsPlugin"
 import NodePolyfillPlugin from "node-polyfill-webpack-plugin"
-import webpack from "webpack"
+import rspack from "@rspack/core"
 
 export const sharedPlugins = () => [
   new NodePolyfillPlugin(),
 
-  new webpack.DefinePlugin({
+  new rspack.DefinePlugin({
     "process.env": {
       NODE_ENV: JSON.stringify(process.env.NODE_ENV),
     },
   }),
-  new webpack.IgnorePlugin({
+  new rspack.IgnorePlugin({
     resourceRegExp: /^\.\/locale$/,
     contextRegExp: /moment$/,
   }),
   // Remove server-only modules from client bundles
   // TODO: Why would these end up in the client bundle?
-  new webpack.IgnorePlugin({ resourceRegExp: /^graphql(\/.*)?$/ }),
+  new rspack.IgnorePlugin({ resourceRegExp: /^graphql(\/.*)?$/ }),
 
   new RetryChunkLoadPlugin({
     maxRetries: 5,