⚠️ 

The type of this PR is: **Feat**

### Description

This migrates our compiler from Webpack to [RSPack](https://rspack.dev/), a webpack-compatable bundler written in Rust for nearly instant cache-free compiles and hotreloading. This should speed up prod deploys considerably while also bringing instant DX.

https://github.com/user-attachments/assets/7d092d18-dd88-41f4-a63f-0ca88c12ecb4

**Cold boot dev compile times:**

Old:
73.31s.

New: 
2.42s

**Prod compile times:** 
Old: 
81.68s

New: 
8.60s

**JS Build Size Comparison:** 

**Dev:**
Webpack: 142mb 
RSPack: 82mb

**Prod**
Webpack: 42mb 
RSPack: 44mb 
RSBuild: 39mb